+++ data_warehouse/forklift/pipeline/flows/clean_flow_runs.py
@@ -3,8 +3,11 @@ from pathlib import Path
 from prefect import Flow, Parameter
 
 from config import FLOW_STATES_TO_CLEAN, MAX_FLOW_RUN_MINUTES
-from src.pipeline.shared_tasks.control_flow import cancel_flow_run, get_flow_runs
-from src.pipeline.shared_tasks.dates import get_timezone_aware_utcnow, make_timedelta
+from forklift.pipeline.shared_tasks.control_flow import cancel_flow_run, get_flow_runs
+from forklift.pipeline.shared_tasks.dates import (
+    get_timezone_aware_utcnow,
+    make_timedelta,
+)
 
 with Flow("Clean flow runs") as flow:
     flow_states_to_clean = Parameter(
