+++ data_warehouse/forklift/pipeline/shared_tasks/dates.py
@@ -4,7 +4,7 @@ from typing import List
 import pytz
 from prefect import task
 
-from src.pipeline.helpers import dates
+from forklift.pipeline.helpers import dates
 
 
 @task(checkpoint=False)
@@ -38,17 +38,18 @@ def make_periods(
     chunk_overlap_minutes: int,
 ) -> List[dates.Period]:
     """
-    `prefect.Task` version of the function `src.pipeline.helpers.dates.make_periods`,
-    with the difference that start and end dates are to be given as a number of hours
-    from the current date (instead of `datetime` objects), and chunk duration and
-    overlap are to be given as a number of minutes (instead of `timedelta` objects).
+    `prefect.Task` version of the function
+    `forklift.pipeline.helpers.dates.make_periods`,with the difference that start and
+    end dates are to be given as a number of hours from the current date (instead of
+    `datetime` objects), and chunk duration and overlap are to be given as a number of
+    minutes (instead of `timedelta` objects).
     This is to accomodate for the fact that Prefect flows' parameters must be
     JSON-serializable, and `datetime` and `timedelta` are not, by default.
 
-    `src.pipeline.helpers.dates.make_periods` is recursive, hence the construction as a
-    python function first and not directly as Prefect `Task`.
+    `forklift.pipeline.helpers.dates.make_periods` is recursive, hence the construction
+    as a python function first and not directly as Prefect `Task`.
 
-    See `src.pipeline.helpers.dates.make_periods` for help.
+    See `forklift.pipeline.helpers.dates.make_periods` for help.
     """
 
     now = datetime.utcnow()
